Tests: CreatePlayer/GetPlayerById (+ BL changes because of the cache options in tests)

+++ EntertainmentPortal/EP.Sudoku.Logic/Queries/GetPlayerById.cs
@@ -5,9 +5,9 @@ namespace EP.Sudoku.Logic.Queries
 {
     public class GetPlayerById : IRequest<Player>
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public GetPlayerById(int id)
+        public GetPlayerById(long id)
         {
             Id = id;
         }
